refactor(sdk): move ABIs from contracts/ to abi/ as single source of truth
Extract inline ABI constants from contract config files into dedicated
abi/*.abi.ts files. Contract files now import ABIs instead of defining
them inline. Remove ABI re-exports from contracts/index.ts since they
are not part of the public API. Delete dead erc20.abi.ts and
batch-swap.abi.ts (unused).
