+++ packages/sdk/src/contracts/acl.ts
@@ -1,39 +1,5 @@
 import type { Address } from "viem";
-
-export const aclAbi = [
-  {
-    inputs: [
-      { internalType: "address", name: "delegate", type: "address" },
-      { internalType: "address", name: "contractAddress", type: "address" },
-      { internalType: "uint64", name: "expirationDate", type: "uint64" },
-    ],
-    name: "delegateForUserDecryption",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "address", name: "delegate", type: "address" },
-      { internalType: "address", name: "contractAddress", type: "address" },
-    ],
-    name: "revokeDelegationForUserDecryption",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "address", name: "delegator", type: "address" },
-      { internalType: "address", name: "delegate", type: "address" },
-      { internalType: "address", name: "contractAddress", type: "address" },
-    ],
-    name: "getUserDecryptionDelegationExpirationDate",
-    outputs: [{ internalType: "uint64", name: "", type: "uint64" }],
-    stateMutability: "view",
-    type: "function",
-  },
-] as const;
+import { aclAbi } from "../abi/acl.abi";
 
 /**
  * Returns the contract config to delegate user decryption rights.
