+++ packages/sdk/src/contracts/erc165.ts
@@ -1,14 +1,5 @@
 import type { Address } from "viem";
-
-export const erc165Abi = [
-  {
-    inputs: [{ internalType: "bytes4", name: "interfaceId", type: "bytes4" }],
-    name: "supportsInterface",
-    outputs: [{ internalType: "bool", name: "", type: "bool" }],
-    stateMutability: "view",
-    type: "function",
-  },
-] as const;
+import { erc165Abi } from "../abi/erc165.abi";
 
 /** ERC-165 interface ID for IERC7984 (confidential fungible token). */
 export const ERC7984_INTERFACE_ID = "0x4958f2a4" as const;
