+++ packages/sdk/src/contracts/index.ts
@@ -1,5 +1,4 @@
 export {
-  encryptedAbi,
   confidentialBalanceOfContract,
   confidentialTransferContract,
   confidentialTransferFromContract,
@@ -16,7 +15,6 @@ export {
 } from "./encrypted";
 
 export {
-  wrapperAbi,
   finalizeUnwrapContract,
   underlyingContract,
   wrapContract,
@@ -32,14 +30,9 @@ export {
   approveContract,
 } from "./erc20";
 
-export {
-  deploymentCoordinatorAbi,
-  getWrapperContract,
-  wrapperExistsContract,
-} from "./deployment-coordinator";
+export { getWrapperContract, wrapperExistsContract } from "./deployment-coordinator";
 
 export {
-  erc165Abi,
   ERC7984_INTERFACE_ID,
   ERC7984_WRAPPER_INTERFACE_ID,
   supportsInterfaceContract,
@@ -48,21 +41,15 @@ export {
 } from "./erc165";
 
 export {
-  feeManagerAbi,
   getWrapFeeContract,
   getUnwrapFeeContract,
   getBatchTransferFeeContract,
   getFeeRecipientContract,
 } from "./fee-manager";
 
-export {
-  transferBatcherAbi,
-  confidentialBatchTransferContract,
-  type BatchTransferData,
-} from "./transfer-batcher";
+export { confidentialBatchTransferContract, type BatchTransferData } from "./transfer-batcher";
 
 export {
-  aclAbi,
   delegateForUserDecryptionContract,
   revokeDelegationContract,
   getDelegationExpiryContract,
