+++ packages/sdk/src/contracts/transfer-batcher.ts
@@ -1,53 +1,6 @@
-import type { Handle } from "../relayer/relayer-sdk.types";
 import type { Address, Hex } from "viem";
-
-export const transferBatcherAbi = [
-  {
-    inputs: [
-      {
-        internalType: "contract RegulatedERC7984Upgradeable",
-        name: "cToken",
-        type: "address",
-      },
-      {
-        internalType: "address",
-        name: "from",
-        type: "address",
-      },
-      {
-        components: [
-          {
-            internalType: "address",
-            name: "to",
-            type: "address",
-          },
-          {
-            internalType: "externalEuint64",
-            name: "encryptedAmount",
-            type: "bytes32",
-          },
-          {
-            internalType: "bytes",
-            name: "inputProof",
-            type: "bytes",
-          },
-          {
-            internalType: "uint256",
-            name: "retryFor",
-            type: "uint256",
-          },
-        ],
-        internalType: "struct ERC7984TransferBatcher.ConfidentialTransferInput[]",
-        name: "transfers",
-        type: "tuple[]",
-      },
-    ],
-    name: "confidentialBatchTransfer",
-    outputs: [],
-    stateMutability: "payable",
-    type: "function",
-  },
-] as const;
+import { transferBatcherAbi } from "../abi/transfer-batch.abi";
+import type { Handle } from "../relayer/relayer-sdk.types";
 
 /** Batch transfer data for confidentialBatchTransfer. */
 export interface BatchTransferData {
