+++ packages/sdk/src/contracts/wrapper.ts
@@ -1,80 +1,6 @@
 import type { Address, Hex } from "viem";
 import type { Handle } from "../relayer/relayer-sdk.types";
-
-export const wrapperAbi = [
-  {
-    inputs: [
-      {
-        internalType: "euint64",
-        name: "burntAmount",
-        type: "bytes32",
-      },
-      {
-        internalType: "uint64",
-        name: "burntAmountCleartext",
-        type: "uint64",
-      },
-      {
-        internalType: "bytes",
-        name: "decryptionProof",
-        type: "bytes",
-      },
-    ],
-    name: "finalizeUnwrap",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "underlying",
-    outputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "to",
-        type: "address",
-      },
-      {
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-    ],
-    name: "wrap",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "to",
-        type: "address",
-      },
-      {
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-    ],
-    name: "wrapETH",
-    outputs: [],
-    stateMutability: "payable",
-    type: "function",
-  },
-] as const;
+import { wrapperAbi } from "../abi/wrapper.abi";
 
 /**
  * Returns the contract config for finalizing an unwrap.
